requirements.txt
django
django-crispy-forms
reportlab
django_otp
djangorestframework-jsonapi